Fix post navigation showing Projects page
For the navigation bar found below each post, the first post would
direct the 'back' navigation to the 'Projects' page. A condition has
been added to check whether or not the page in question is the index.md
page of /projects before referring.

Also fixed an issue with the CSS where it would make the entire side of
the posts navigation bar a link, rather than just the text.